+++ src/main/java/kitchenpos/product/domain/Price.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.product.domain;
 
 import javax.persistence.Embeddable;
 import java.math.BigDecimal;
